Removed old matcher implementation class
